// hivemq-edge-frontend/src/api/__generated__/models/FunctionSpecsList.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FunctionSpecs } from './FunctionSpecs';

/**
 * List of function configurations
 */
export type FunctionSpecsList = {
    /**
     * List of function configurations
     */
    items: Array<FunctionSpecs>;
};