// hivemq-edge-frontend/src/api/__generated__/models/BehaviorPolicyTransitionEvent.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Accepted event in transition
 */
export enum BehaviorPolicyTransitionEvent {
    EVENT_ON_ANY = 'Event.OnAny',
    CONNECTION_ON_DISCONNECT = 'Connection.OnDisconnect',
    MQTT_ON_INBOUND_CONNECT = 'Mqtt.OnInboundConnect',
    MQTT_ON_INBOUND_DISCONNECT = 'Mqtt.OnInboundDisconnect',
    MQTT_ON_INBOUND_PUBLISH = 'Mqtt.OnInboundPublish',
    MQTT_ON_INBOUND_SUBSCRIBE = 'Mqtt.OnInboundSubscribe',
}